requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==36.0.2
disnake==2.5.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
kaitaistruct==0.9
multidict==6.0.2
outcome==1.1.0
pkg_resources==0.0.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.27.1
selenium==4.1.3
selenium-wire==4.6.3
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.9
wsproto==1.1.0
yarl==1.7.2
zstandard==0.17.0
